added navbar component
