+++ app/javascript/components/pages/Home.js
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import SkillTrackerButton from "../elements/button";
 
 const Home = () => {
